Renamed files and references to Zotapay
