+++ package.json
@@ -2,21 +2,23 @@
   "name": "@zota/sdk",
   "author": "Zota Technology Pte LTD",
   "version": "1.3.1",
-  "description": "This Node.js SDK provides all the necessary methods for integrating the Zotapay Merchant API.",
+  "description": "This Node.js SDK provides the necessary methods for integrating Zota's Merchant API.",
   "repository": {
     "type": "git",
-    "url": "https://github.com/zotapay/node-sdk.git"
+    "url": "https://github.com/zota/node-sdk.git"
   },
   "bugs": {
-    "url": "https://github.com/zotapay/node-sdk/issues"
+    "url": "https://github.com/zota/node-sdk/issues"
   },
-  "homepage": "https://github.com/zotapay/node-sdk#readme",
+  "homepage": "https://github.com/zota/node-sdk#readme",
   "keywords": [
     "zotapay",
+    "zota",
     "node.js",
     "zota.com",
     "payments",
     "alternative-payments",
+    "alternative-payment-methods",
     "sdk"
   ],
   "license": "Apache-2.0",
