+++ zotasdk/helpers.js
@@ -1,6 +1,6 @@
 const HTTP_STATUS_OK = 200
 
-const UA_STRING = "Zotapay/NodeJS-SDK %s ( %s) NodeJS ( %s)"
+const UA_STRING = "Zota/NodeJS-SDK %s (%s) NodeJS (%s)"
 
 module.exports = {
     HTTP_STATUS_OK: HTTP_STATUS_OK,
